server/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
ffmpeg-python==0.2.0
future==1.0.0
gunicorn==25.3.0
packaging==26.0
pillow==12.2.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
tzdata==2026.1
